requirements.txt
flask
joblib
numpy
langchain
langchain-google-genai
pydantic
python-dotenv
google-generativeai